v0.9

+++ src/MakeBloomFilter.cpp
@@ -41,7 +41,7 @@ BF<LARGE_BITSET> MakeBF(ReadSet &RS,KmerCount &KC,uint64_t filtersize ,uint8_t n
     std::vector<LARGE_BITSET> end_bases={A_left,C_left,G_left,T_left,A_right,C_right,G_right,T_right};
 
     const uint32_t shortk_length=21;
-    const uint32_t cov_threshold=3;
+    const uint64_t cov_threshold=3;
 
     #pragma omp parallel for  num_threads(20) 
     for(size_t b=0;b<RS.bucket_count();b++)
